oops - use existing test_publish file for new test; fix noninteractive publish
